# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EntEx API",
    "version": "1.0.0",
    "description": "Multi-tenant platform for entertainment booking, rights management, agent marketplace, and double-entry ledger."
  },
  "servers": [{ "url": "http://localhost:3000/api/v1", "description": "Local development" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "tenantId": { "type": "apiKey", "in": "header", "name": "X-Tenant-ID" }
    },
...
  },
  "paths": {
...
    "/businesses": {
      "get": {
        "summary": "List businesses for tenant",
        "tags": ["Business"],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Paginated business list" }
        }
      },
      "post": {
        "summary": "Create a business",
        "tags": ["Business"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "vertical": { "type": "string", "default": "entertainment" },
                  "legalName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Business created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/businesses/{id}": {
      "get": {
        "summary": "Get business by ID",
        "tags": ["Business"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Business details" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update business",
        "tags": ["Business"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Business updated" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Archive business",
        "tags": ["Business"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Business archived" }
        }
      }
    },
...
  },
...
}